backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Database & Firebase
firebase-admin==6.4.0
pymongo==4.6.0

# AI/ML & Speech Processing
openai==1.3.5
speech-recognition==3.10.0
pyttsx3==2.90
transformers==4.36.0
torch==2.1.0
opencv-python==4.8.1.78

# Audio Processing
librosa==0.10.1
soundfile==0.12.1

# Image Processing
Pillow==10.1.0
numpy==1.25.2

# HTTP & API
requests==2.31.0
aiohttp==3.9.0
httpx==0.25.2

# Authentication & Security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
bcrypt==4.1.1

# Utilities
python-dotenv==1.0.0
asyncio==3.4.3
typing-extensions==4.8.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
